pyproject.toml
[project]
name = "chess-autoresearch"
version = "0.1.0"
description = "Autonomous chess engine research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-chess>=1.999",
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "pandas>=2.3.0",
]
